DWPDAT-25: add package for benefit data retrival classes
