+++ app/uk/gov/hmrc/nationalinsurancecontributionandcreditsapi/services/benefitEligibilityDataRetrieval/BenefitEligibilityDataRetrievalService.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2025 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.services.benefitEligibilityDataRetrieval
+
+import cats.data.EitherT
+import uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.models.api.EligibilityCheckDataRequest
+import uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.models.domain.EligibilityCheckDataResult
+
+import scala.concurrent.Future
+
+trait BenefitEligibilityDataRetrievalService {
+
+  // TODO Return a custom error instead of Throwable once the error schema to be returned to DWP is decided
+  def fetchEligibilityData(
+      eligibilityCheckDataRequest: EligibilityCheckDataRequest
+  ): EitherT[Future, Throwable, EligibilityCheckDataResult]
+
+}
